Added program to implement custom linked list and refactored linkedlist folder
