+++ Programs/LinkedList/Circular Integer LinkedList/Node.java
@@ -1,0 +1,19 @@
+public class Node {
+	int data;
+	Node next;
+	
+	public Node(int a) {
+		data=a;
+		next=null;
+		
+	}
+	
+	public Node() {
+		data=-1;
+		next=null;
+	}
+
+	public int getNodeElement(){
+		return data;
+	}
+}
